firefox_container/requirements.txt
selenium
google-cloud-storage
pygsheets
pandas
numpy
httplib2shim